Update README.md
Hi Everyone!

i created a weather app using html,css,javascript.



You must know the basics in JavaScript and HTML, like JavaScript Objects, functions, If statements, ... etc

The only difficulty i face is how to work with an API using JavaScript? but trust me that won't be of a big problem.

Let's talk about the application you're going to build; our weather app will show the weather information that it will get from the API like this, an Icon that shows the weather status, the temperature value (33°C) in Celsius unit, the weather description (broken clouds), and then the user's city and country (raipur, IN).

When the user clicks on the temperature value, that will convert it from Celsius to Fahrenheit.